pyproject.toml
[tool.poetry]
name = "impulseed-autotests"
version = "0.1.0"
description = ""
authors = ["Aleksandr Dymchenko <av.dymchenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.4"
httpx = "^0.28.1"
allure-pytest = "^2.13.5"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
